Update README.md
Welcome to the ProjectLibre community.  
ProjectLibre was voted open source 'Project of the Month' in our first month!!!!! 
Thank you to the community,  it is important as we are looking to grow 
the usage of our open source replacement of Microsoft Project 
with a vibrant international community.  
You can join the community today! The ProjectLibre team were 
the founders and developers of open source software downloaded
more than 4 million times around the world. 
ProjectLibre is a major open source addition as the 
leading alternative to Microsoft Project.
It has been downloaded in 192 countries Download Now!!!!</a>


Naročnik: <a href="https://plus.google.com/102962854949157079552">
ZekomZ - Zekom Tribute Magazine - G+
</a>
<a href="http://www.indiegogo.com/zekom">WwW.IndieGoGo.com/Zekom</a>
Zekom. It's the Law.
Sodelujoči:
https://github.com/zekom
https://github.com/lukas86